Another update to ForLife lid
This version reverts to the older diameter for the center body,
which permits to lid to settle into the infuser. The previous
version was made wide enough to center onto the cup, but obviously
the tea is brewed in the infuser, so that was a mistake. However,
in order to retain the capability of covering the cup, brim is
made wider now.