AR9MM: Improve the feed ramp
We were unable to move the magazine, and not getting its follower
or feed lips whacked by the bolt, so the feed ramp was found very
necessary. Unfortunately, the magazine cannot be moved to the rear
either, else the AR magazine catch hits the pistol magazine body.
There's not much material in front.

We also move the ejector a little to the rear and add the lettering.

Prototype: No.4