RU9: Remove the printed ejector
Our simplistic printed ejector is not strong enough, breaks off
immediately. The AR9LS goes with plastic, but they are disposable.
So, let's just cut the losses.

This is still an experimental revision, so it does not include
the drawings for the metal parts. They are not finalized yet.

Prototype: P.04