Improve Kuhn-Rikon deflector core some more
We broke off one of the petals in the previous version.
But making it thicker won't do. So, we increased the
inside diameter by 0.4 mm, in order to reduce the strain
on the petals. The previous version slid onto the valve
with an undue friction anyway. Also, unfortunately, we had
to reduce the height of the petals by 1 mm. In order to
compensate for it somewhat, and to make the ring less
rigid, we made the cuts a little deeper.