Add the scalable mushroom form
This parametrizes the forms, in order to make them bigger easier.
But for now, this produces exactly the same mushroom as before.

Because we made everything parametrized, it was very easy to
create an almost-minkovsky shape. No point to waste the filament
and printing time to a bunch of cylinders.

Because the body is now mostly smooth, we added zip-tie stops.